feat: verify that the file exists

+++ src/application/utils/FileUtils.java
@@ -23,15 +23,41 @@ public class FileUtils {
      * @param pathURI - the path to create the directory.
      * @return true if its created, false otherwise.
      */
-    public boolean create(String pathURI) {
+    public boolean createDirectory(String pathURI) {
         File f = new File(pathURI);
         if(f.isFile()) return false;
         if(f.isDirectory() && f.exists()) return true;
-        if(f.toPath().getNameCount() > 2) {
-            return f.mkdirs();
-        }
+        if(f.toPath().getNameCount() > 2) return f.mkdirs();
         return f.mkdir();
-
+    }
+    /**
+     * Deletes a directory if its empty, otherwise you must provide --r to delete it.
+     * @param pathURI - the directory to delete.
+     * @param recursively - to delete a directory that isn't empty.
+     */
+    public boolean deleteDirectory(String pathURI, boolean recursively) {
+        File f = new File(pathURI);
+        if(!f.isDirectory() || !f.exists()) return false;
+        try {
+            if(recursively) {
+                // get files to delete.
+                List<Path> paths = listDirContent(pathURI, 0)
+                    .stream()
+                    .filter(Files::isRegularFile)
+                    .toList();
+                for(Path p: paths) {
+                    System.out.println(String.format("[Info] Deleting => %s", p.toString()));
+                    Files.deleteIfExists(p);
+                }
+                return deleteDirectory(f.getPath(), false);
+            } else {
+                System.out.println(String.format("[Info] Deleting => %s", f.toString()));
+                return Files.deleteIfExists(f.toPath());
+            }
+        } catch(IOException e) {
+            e.printStackTrace();
+            return false;
+        }
     }
     /**
      * List directory content by level, if nested level is 0, it will search recursively.
@@ -54,12 +80,12 @@ public class FileUtils {
      * @param targetURI - the path where to copy the source file.
      */
     public void copyFileToTarget(Path sourcePath, String targetURI) {
-        if(!sourcePath.toFile().isFile()) return;
+        if(!sourcePath.toFile().isFile() || !sourcePath.toFile().exists()) return;
         try {
             Path destination = Paths.get(targetURI).resolve(sourcePath.getFileName());
             Path result = Files.copy(sourcePath, destination, StandardCopyOption.COPY_ATTRIBUTES);
             if(result != null) {
-                System.out.println(String.format("[Info] copy %s into \n\t=>[%s]", sourcePath.toString(), destination.toString()));
+                System.out.println(String.format("[Info] copy %s \n\tinto \t=>[%s]", sourcePath.toString(), destination.toString()));
             }
         } catch(IOException e) {
             e.printStackTrace();
@@ -72,7 +98,7 @@ public class FileUtils {
      * @param level - the nested level to reach.
      */
     public void copyDirToTarget(Path sourcePath, String targetURI, int level) {
-        if(!sourcePath.toFile().isDirectory()) return;
+        if(!sourcePath.toFile().isDirectory() || !sourcePath.toFile().exists()) return;
         try {
             Path targetPath = Paths.get(targetURI);
             // first list and create the directory structure.
@@ -81,11 +107,11 @@ public class FileUtils {
                 Path relative = sourcePath.relativize(p);
                 Path destination = targetPath.resolve(relative);
                 if(Files.isDirectory(p)) {
-                    create(destination.toString());
+                    createDirectory(destination.toString());
                     System.out.println(String.format("[Info] Creating {%s}", destination.toString()));
                 } else {
                     Path r = Files.copy(p, destination, StandardCopyOption.COPY_ATTRIBUTES);
-                    System.out.println(String.format("[Info] Copy %s into \n\t=>[%s]", p, r));
+                    System.out.println(String.format("[Info] Copy %s \n\tinto \t=>[%s]", p, r));
                 }
             }
         } catch(Exception e) {
